oauth2_cli_github_example
Module encrypt_decrypt_with_ssh_key_mod

encrypt_decrypt_with_ssh_key_mod

Functions to encrypt decrypt a secret string

Structs

EncryptedTextWithMetadata 🔒

Struct that represents the json data saved in the ‘*.enc’ file

Constants

BLUE

ANSI color

GREEN

ANSI color

RED

ANSI color

RESET

ANSI color

YELLOW

ANSI color

Functions

decode64_from_string_to_32bytes 🔒

Decode base64 from string to 32bytes

decode64_from_string_to_string 🔒

Decode base64 from string to string It is a silly little obfuscation just to avoid using plain text.

decrypt_symmetric 🔒

Decrypts plain_encrypted_string with secret_passcode_32bytes

encode64_from_32bytes_to_string 🔒

Decode base64 from string to 32bytes

encode64_from_bytes_to_string 🔒

Encode base64 from bytes to string

encode64_from_string_to_string 🔒

Encode base64 from string to string It is a silly little obfuscation just to avoid using plain text.

encrypt_symmetric 🔒

Encrypts symmetrically secret_string_to_encrypt with secret_passcode_32bytes

open_file_get_string 🔒

get the string from the file that is base64 encoded it is encoded just to obscure it a little

random_seed_32bytes 🔒

Generate a random seed

shorten_vec_bytes_to_32bytes 🔒

shorten the Vec<u8> to [u8;32]

sign_seed_with_identity_file 🔒

Sign the seed with the private key into 32 bytes secret

sign_seed_with_ssh_agent 🔒

Sign seed with ssh-agent into 32 bytes secret

sign_seed_with_ssh_agent_or_identity_file 🔒

Returns the signed seed. That will be the actual password for symmetrical encryption. Returns secret_password_bytes