oauth2_cli_github_example
Module github_oauth2_device_workflow_mod

Constants

BLUE

ANSI color

GREEN

ANSI color

RED

In this module there will be a lot of work with secrets. It is difficult to trust an external crate with your secrets. The crates can get updated unexpectedly and change to malicious code. It is best to have the Rust code under your fingertips when dealing with secrets. Than you know nobody will touch this code except of you. You can copy this code directly into your codebase, inspect and review it and know exactly what is going on. ANSI color

RESET

ANSI color

YELLOW

ANSI color

Functions

first_authentication_with_browser 🔒

github_oauth2_device_workflow 🔒