oauth2_cli_github_example
Module encrypt_decrypt_with_ssh_key_mod

generic functions to encrypt and decrypt secrets using the ssh private key

Functions to encrypt decrypt a secret string

Modules

crates_io_api_token_mod

decrypt crates.io api token from file or ask the user to input the access token, encrypt it and save into file

docker_io_api_token_mod

decrypt docker.io api token from file or ask the user to input the access token, encrypt it and save into file

generate_strong_password_mod

github_api_token_with_oauth2_mod

decrypt github api token from file or use the oauth2 device workflow to get the access token and encrypt it and save into file

secret_vault_mod

small secret vault CLI

Structs

EncryptedTextWithMetadata 🔒

Struct that represents the json data saved in the ‘*.enc’ file

Constants

BLUE

ANSI color

GREEN

ANSI color

RED

ANSI color

RESET

ANSI color

YELLOW

ANSI color

Functions

decode64_from_string_to_32bytes 🔒

Decode base64 from string to 32bytes

decode64_from_string_to_string 🔒

Decode base64 from string to string

decrypt_symmetric 🔒

Decrypts plain_encrypted_string with secret_passcode_32bytes

encode64_from_32bytes_to_string 🔒

Decode base64 from string to 32bytes

encode64_from_bytes_to_string 🔒

Encode base64 from bytes to string

encode64_from_string_to_string 🔒

Encode base64 from string to string

encrypt_symmetric 🔒

Encrypts symmetrically secret_string_to_encrypt with secret_passcode_32bytes

open_file_b64_get_string 🔒

Get the string from the file that is base64 encoded

random_seed_32bytes_and_string 🔒

Generate a random seed

shorten_vec_bytes_to_32bytes 🔒

shorten the Vec<u8> to [u8;32]

sign_seed_with_private_key_file 🔒

Sign the seed with the private key into 32 bytes secret

sign_seed_with_ssh_agent 🔒

Sign seed with ssh-agent into 32 bytes secret

sign_seed_with_ssh_agent_or_private_key_file 🔒

Returns the secret signed seed